==> 02-data.Rmd <==
# Data sources

**Note:** For this project, data was extracted on around November 20, 2020. There was a significant update to the database sometime around December 16, 2020 with the release of the 2020 Human Development Report. Changes include, but were not limited to, adding 2019 human development data, updating computation methodology of certain indices (which caused slight changes to pre-existing values), and adding/removing features for certain indicators. This report is consistent with the state of the database prior to the update, and where applicable, hyperlinks in the follow sections direct to the archived version of the pages (as of November 24-25, 2020) referenced by this report.  

The data used for this project is downloaded directly from the [United Nations Development Programme - Human Development Reports website](https://web.archive.org/web/20201124153155/http://hdr.undp.org/en/data). As noted above, this links to the archived version of the website used for this report and will take a few minutes to load. The live version of this website is [here](http://hdr.undp.org/en/data). 

The United Nations Development Programme's (UNDP) Human Development Report Office (HDRO) compiles this data for the purpose of tracking human wellbeing in countries around the world. This data is used extensively by the HDRO in its (near) annual Human Development Report (HDR). The report highlights human development in large part using various human development indices (HDI), derived metrics for measuring human development, which take into account factors beyond economics. However, for this project HDI will not be a principal part of the investigation.

According to the [Human Development Report 2019: Reader's Guide](https://web.archive.org/web/20201125185918/http://hdr.undp.org/en/content/human-development-report-2019-readers-guide), the data used, at least for the 2019 HDR, is collected "from international data agencies with the mandate, resources, and expertise to collect national data on specific indicators." Based on the "Statistical Acknowledgment" section of the guide, the HDR data was aggregated by the HDRO from a large number of organizations that are reputable and specializes in the field from which the data was collected. For the 2019 HDR, these included international agencies like [the World Health Organization, the World Bank, and many others](http://hdr.undp.org/sites/default/files/2019_statistical_references.pdf). The HDRO also provides [contact information](http://hdr.undp.org/en/statistics/understanding/sources), where available, for organizations it sourced HDR data from.           

The data available for download from UNDP's website linked in the first paragraph consists of 12 different indicators of human development (plus HDI) and for each indicator, there are a number of descriptive features of that indicator for which data is available. For example, there are 9 features which fall under the `Demography` indicator, ranging from "Median age (years)" to "Young age (0-14) dependency ratio (per 100 people ages 15-64)". The data for each feature in every indicator was downloaded by first filtering the "Dimension" drop down menu, then by filtering for the relevant feature in a second drop down menu, and lastly going to "Download Data" to save the feature data as a .csv file. For each feature, a data table, if available, is provided where the columns include "HDI Rank (2018)", "Country", and one column for every year (i.e. 1990) for which data is aggregated. Every column except "Country" contains numerical data. The years for which data is available falls in the range from 1990 to 2018.       

For this project, we will be examining 3 of the 12 indicators: Inequality, Human Security, and Education, which have 15, 7, and 22 features respectively. These features will be investigated in further detail in subsequent sections.     

Investigation showed that there are a number of issues with this dataset, as summarized below:

* The [Human Development Report 2019: Reader's Guide](https://web.archive.org/web/20201125185918/http://hdr.undp.org/en/content/human-development-report-2019-readers-guide) note that due to data accessibility and processing methodology, there may be discrepancies in data between national and international agencies.

* While the dataset contains data from 1990 to 2018, it does not consistently store data for this timeframe for every feature of every indicator. For example, some features have data going back to 1990 but others only back to 2010.

* Even within the same indicator, different features show data across different time intervals. For example, some features have data for each year, others every 5 years. 

* Some features do not have data in even time intervals. For example, "Prison population (per 100,000 people)" has data for years 2003, 2005, 2010, 2011 etc.  

* As detailed in subsequent sections, the proportion of missing data is high for some countries. 

* For certain feature/country combinations, 2018 data was not available and was filled with the data from the most recent available year.

* It was not explicitly stated in the source documentation, but given that the data is collected from 1990-2018 and aggregated from numerous sources, it is possible that there lacks consistency in data collection and computation of certain metrics over time. [For instance, method of calculating the HDI is continuously being adjusted.](http://hdr.undp.org/en/humandev)

Additionally, the process of international data collection provides its own set of challenges. The [Reader's Guide](https://web.archive.org/web/20201125185918/http://hdr.undp.org/en/content/human-development-report-2019-readers-guide) notes that the data are collected from "international data agencies with the mandate, resources, and expertise to collect national data on specific indicators," but also (as above) notes the potential for discrepancies. International data collection can be complicated by numerous factors, including variation in infrastructural development, access to technology or stable internet connection, incidence of human rights or humanitarian crises, political violence or instability, and more. As this study focuses on patterns among factor variables, rather than patterns in missing or incomplete country data, we do not specifically address these potential impediments. We posit this trend as an important direction for further research.
