==> 03-cleaning.Rmd <==
# Data transformation

The data transformation work was split into three separate buckets, one for each indicator: Inequality, Human Security, and Education. Data transformation for every indicator was similar to the process shown in detail for Inequality, which is not fully reproduced for the other two indicators for conciseness.


## Inequality Data Transformation

The Inequality indicator has 15 features from years 2010 to 2018 (one column per year). However, the data table for 6 of these features were not available for download. The remaining 9 features for which data tables were available and was extracted were:

* Coefficient of Human Inequality

* Inequality-adjusted education index

* Inequality-adjusted HDI (IHDI)

* Inequality-adjusted income index

* Inequality-adjusted life expectancy index

* Inequality in education (%)^

* Inequality in income (%)^

* Inequality in life expectancy (%)

* Overall loss in HDI due to inequality (%)

^ Data for 2018 is either from 2018 or the latest year available. 

Note that the column "HDI Rank (2018)" is kept as a reference column and not for analysis. First, all of the files are read by parsing through the Inequality/inequality_rawdata directory, and using read_csv to read in the relevant raw data files for each variable. The files are read in iteratively such that the data frame for every feature is stored in a R list, creating a list of data frames. 

```{r}
library(tidyverse)
library(readxl)
library(readr)
library(stringr)
```

```{r warning = FALSE,  message = FALSE}
temp <- list.files(path = "./Inequality/inequality_rawdata", pattern="*.csv")
 
inequality_list <- list()
for(i in 1:length(temp)){
  inequality_list[[i]] <- read_csv(paste0("./Inequality/inequality_rawdata/", temp[i]), skip = 1) %>%
                    select("HDI Rank (2018)", Country, `2010`, `2011`, `2012`, `2013`, 
                           `2014`, `2015`, `2016`, `2017`, `2018`) 
}
names(inequality_list) <- sub("\\..*", "", temp)
```

After reading in the files, a high level look at the data itself was done. The head and tail of one of the feature data frames, containing data for "Inequality in education (%)", was printed as a representative sample:

```{r}
print(head(inequality_list[[6]]))
print(tail(inequality_list[[6]]))
```

Inspection of the remaining data frame indicate that the last rows for the data frames are every feature are NA. In addition, some data frames have foot notes at the end of the table that was read in as a value in the first column. These rows are removed. 

```{r}
inequality_list[[1]] <- inequality_list[[1]] %>% 
                            filter(row_number() <= n()-1)
inequality_list[[2]] <- inequality_list[[2]] %>% 
                            filter(row_number() <= n()-1)
inequality_list[[3]] <- inequality_list[[3]] %>% 
                            filter(row_number() <= n()-1)
inequality_list[[4]] <- inequality_list[[4]] %>% 
                            filter(row_number() <= n()-1)
inequality_list[[5]] <- inequality_list[[5]] %>% 
                            filter(row_number() <= n()-1)
inequality_list[[6]] <- inequality_list[[6]] %>% 
                            filter(row_number() <= n()-3)
inequality_list[[7]] <- inequality_list[[7]] %>% 
                            filter(row_number() <= n()-3)
inequality_list[[8]] <- inequality_list[[8]] %>% 
                            filter(row_number() <= n()-2)
inequality_list[[9]] <- inequality_list[[9]] %>% 
                            filter(row_number() <= n()-1)
```

The output below shows that the number of rows in the data frame for each feature is different. This implies not every feature has the same list of countries. 

```{r warning = FALSE,  message = FALSE}
for(i in 1:length(inequality_list)){
  row <- dim(inequality_list[[i]])[1]
  column <- dim(inequality_list[[i]])[2]
  print(paste0(sub("\\..*", "", temp[i])," feature has ", row, " rows and ", column, " columns."))
}
```
To investigate this further:

```{r, message = FALSE, warning = FALSE}
differences <- c()
for(i in 1:length(inequality_list)){
  for(j in 1:length(inequality_list)){
    differences <- c(differences,setdiff(inequality_list[[i]]$Country, inequality_list[[j]]$Country))
  }
}
print(unique(differences))
```
It looks like these four countries shows up for some features and not others. We can expect a lot of missing data for at least these four countries. 

The next step is to create a single tidy data frame for Inequality. The data frame for each feature has column headings of 2010 to 2018, which are values, not variables. In addition, each data frame is from the same observational unit (Inequality) and should be in one data frame. To convert the list of data frames created from the initial import from the .csv files into a single tidy data frame, use `pivot_longer` and `full_join` to create a tidy data frame.   

```{r}
inequality_df <- inequality_list[[1]] %>% 
                    pivot_longer(!c('HDI Rank (2018)', 'Country'), 
                                 names_to = "Year", 
                                 values_to = names(inequality_list[1]))
for(i in 2:length(inequality_list)){
  temp_df <- inequality_list[[i]] %>% 
                    pivot_longer(!c('HDI Rank (2018)', 'Country'), 
                                 names_to = "Year", 
                                 values_to = names(inequality_list[i]))
  inequality_df <- inequality_df %>%
                      full_join(temp_df %>% select(-'HDI Rank (2018)'), 
                                by = c('Country','Year'))
} 
head(inequality_df)
```

Note that the columns of tidy data frame are currently all of type `<chr>` when only "Country" should be of type `<chr>`, "Year" should be of type integer, and the remaining be of type numeric. This is because missing values are given as '..' instead of `NA`.  Convert the relevant columns to numeric which coerces '..' to `NA`'s in the process.  

```{r warning = FALSE}
inequality_df[,4:12] <- as.data.frame(sapply(inequality_df[,4:12], as.numeric))
inequality_df$Year <- as.integer(inequality_df$Year)
head(inequality_df)
```

To complete the data transformation, notice that under "Country", the last 17 rows of each data frame contain aggregate  regional data, and are not countries: 

```{r}
tail(inequality_list[[1]]$Country,17)
```
Therefore, they will be removed from the tidy data frame. The head and tail of the tidy data frame is:

```{r}
not_countries <- tail(inequality_list[[1]]$Country,17)
 
inequality_df <- inequality_df %>% 
  filter(!(Country %in% not_countries))
```
```{r}
head(inequality_df)
tail(inequality_df)
inequality_df %>% write_csv('./Inequality/inequality_tidy.csv')
```

## Human Security Data Transformation


Below are the variables pulled from the Human Security dataset:

* Homicide Rate (per 100,000 people)

* Prison Population (per 100,000 people)

* Suicide Rate (Male) (per 100,000 people)

* Suicide Rate (Female) (per 100,000 people)

And, as addressed below:

* Homelessness due to Natural Disaster (average annual per million people)

* Refugees by Country of Origin


Transforming this data largely required reorganizing the dataframe into a tidy format.  This was done in a similar fashion to the inequality, and education data manipulations.  This required removing blank columns (which had been left in the .csv files for visual appeal) and renaming columns to prime them for the pivot_longer function. Then, to ensure that all missing data were coded appropriately, a for loop was introduced to convert all numeric data columns in the dataset into numeric type.  The result is a tidy dataset of all Human Security variables.


```{r}

Homeless.natural.disaster <- read.csv("./HumanSecurity/Homeless.natural.disaster.csv", fileEncoding = "UTF-8")
Homicide.rate <- read.csv("./HumanSecurity/Homicide.rate.csv", fileEncoding = "UTF-8")
Prison.popn <- read.csv("./HumanSecurity/Prison.population.csv", fileEncoding = "UTF-8")
Refugees <- read.csv("./HumanSecurity/Refugees.CO.csv", fileEncoding = "UTF-8")
Suicide.rate.M <- read.csv("./HumanSecurity/Suicide.rate.M.csv", fileEncoding = "UTF-8")
Suicide.rate.F <- read.csv("./HumanSecurity/Suicide.rate.F.csv", fileEncoding = "UTF-8")
```

First the blank and extraneous rows are removed from each dataset, and the column names are converted to reflect the year each measurement was taken.

```{r}

names(Homeless.natural.disaster)[2:30] <- c("1990", "1991", "1992", "1993", "1994", "1995", "1996", "1997", "1998", "1999", "2000", "2001", "2002", "2003", "2004", "2005", "2006", "2007", "2008", "2009", "2010", "2011", "2012", "2013", "2014", "2015", "2016", "2017", "2018")

Homicide.rate <- subset(Homicide.rate, select = -c(X, X.1, X.2, X.3, X.4, X.5, X.6, X.7, X.8, X.9, X.10))
names(Homicide.rate)[2:13] <- c("1990", "1995", "2000", "2005", "2010", "2011", "2012", "2013", "2014", "2015", "2016", "2017")

Refugees <- subset(Refugees, select = -c(1))
names(Refugees)[2:30] <- c("1990", "1991", "1992", "1993", "1994", "1995", "1996", "1997", "1998", "1999", "2000", "2001", "2002", "2003", "2004", "2005", "2006", "2007", "2008", "2009", "2010", "2011", "2012", "2013", "2014", "2015", "2016", "2017", "2018")

Prison.popn <- subset(Prison.popn, select = -c(X, X.1, X.2, X.3, X.4, X.5, X.6, X.7, X.8, X.9))
names(Prison.popn)[2:11] <- c("2003", "2005", "2010", "2011", "2012", "2013", "2014", "2015", "2016", "2017")

Suicide.rate.F <- subset(Suicide.rate.F, select = -c(X, X.1, X.2))
names(Suicide.rate.F)[2:5] <- c("2000", "2010", "2015", "2016")

Suicide.rate.M <- subset(Suicide.rate.M, select = -c(X, X.1, X.2))
names(Suicide.rate.M)[2:5] <- c("2000", "2010", "2015", "2016")
```


At this point, a problem became apparent.

Though the UN Data website lists both a Refugees dataframe and a Homelessness due to natural disasters dataframe as within the available datasets, there seemed to be no actual data available for either.  When the Refugees by country of origin and Homelessness due to natural disasters datasets were downloaded (per the option on the UN Data website), the files, though titled correctly, were copies of the HDI records.

Below, the initial row of each dataset is compared to show their similarity.

```{r}
compare <- Refugees[1,1:9]
compare[2,] <- merge(compare, Homeless.natural.disaster[1,1:9])
compare
```

A note was sent to the administrators of the UN Data website to alert them to this issue.

However, as a result, the Refugees by country of origin and Homelessness from natural disaster datasets were excluded from this study.

Next, the numeric data columns are converted to numeric type.  Note that all columns are originally in character type.  Converting to numeric type both eases analysis and allows the blank entries, denoted as ".." to be read as empty (NA).

```{r, warning = FALSE}

dcol <- 2
for (i in 2:ncol(Homicide.rate)) {
  Homicide.rate[,dcol] <- as.numeric(Homicide.rate[,dcol])
  dcol <- dcol +1
}

dcol <- 2
for (i in 2:ncol(Prison.popn)) {
  Prison.popn[,dcol] <- as.numeric(Prison.popn[,dcol])
  dcol <- dcol +1
}

dcol <- 2
for (i in 2:ncol(Suicide.rate.F)) {
  Suicide.rate.F[,dcol] <- as.numeric(Suicide.rate.F[,dcol])
  dcol <- dcol +1
}

dcol <- 2
for (i in 2:ncol(Suicide.rate.M)) {
  Suicide.rate.M[,dcol] <- as.numeric(Suicide.rate.M[,dcol])
  dcol <- dcol +1
}
```

Finally, the data are converted to a tidy format and merged.

```{r}
HR <- pivot_longer(Homicide.rate, cols = !Country, names_to = "Year", values_to = "Homicide.rate")
PP <- pivot_longer(Prison.popn, cols = !Country, names_to = "Year", values_to = "Prison.popn")
SRF <- pivot_longer(Suicide.rate.F, cols = !Country, names_to = "Year", values_to = "Suicide.rate.F")
SRM <- pivot_longer(Suicide.rate.M, cols = !Country, names_to = "Year", values_to = "Suicide.rate.M")

Human.Security <- merge(HR, PP, all.x = T, all.y = T)
Human.Security <- merge(Human.Security, SRF, all.x = T, all.y = T)  
Human.Security <- merge(Human.Security, SRM, all.x = T, all.y = T)
Human.Security <- Human.Security[30:5684,]
```

```{r, eval=F}
Human.Security %>% write_csv('./HumanSecurity/HumanSecurity.Final.Tidy.csv')
```



## Education Data Transformation

The following are the columns in the education dataset:

* Education Index

* Expected years of schooling (years)

* Expected years of schooling, female (years)

* Expected years of schooling, male (years)

* Government expenditure on education (% of GDP)

* Gross enrollment ratio, pre-primary (% of preschool-age children)

* Gross enrollment ratio, primary (% of primary school-age children)

* Gross enrollment ratio, secondary (% of secondary school-age children)

* Gross enrollment ratio, tertiary (% of tertiary school-age children)

* Literacy rate, adult (% ages 15 and older)

* Mean years of schooling (years)

* Mean years of schooling, female(years)

* Mean years of schooling, male (years)

* Percentage of primary schools with access to the internet

* Percentage of secondary schools with access to the internet

* Population with at least some secondary education (% ages 25 and older)

* Population with at least some secondary education, female (% ages 25 and older)

* Population with at least some secondary education, male (% ages 25 and older)

* Primary school dropout rate (% of primary school cohort)

* Primary School teachers trained to teach (%)

* Pupil-teacher ratio, primary school (number of pupils per teacher)

* Survival rate to the last grade of lower secondary general education (%)


Note that the column 'HDI Rank (2018)' is kept as a reference column and not for analysis. 


In order to read in, and transform the education data, a method similar to the previous two situations is used. Every file is found, and read from the relevant Education directory. With this data we determine the relevant NAs in the files, and code those as such. Similarly, we skip the first lines in the files which are blank. Then we remove any rows that are irrelevant-- specifically the rows with footnotes. Finally, we remove additional columns, and change data types if necessary, and are then ready to transform the data into tidy format.

```{r}

dr="./Education/csv"
file_list <- list.files(path=dr,pattern="*.csv")
dttyp<-c("EdInd","ExpSch","ExpSchF","ExpSchM","GovMon","GERPP","GERP","GERS","GERT",
         "LitR","MYSch","MYSchF","MYSchM","PerPSInt","PerSSInt",
         "PopSS25U","PopSSF25U","PopSSM25U",
         "PSDrop","PSTchTrn","PupTchRPS","SurvRSGE")

for (i in 1:length(file_list)){
   df_temp<-read.csv(paste(dr,"/",file_list[i],sep=""),skip=1,header=T, 
                      check.names = F,na.strings=c("..",""," ","<NA>"),
                     fileEncoding = "UTF-8")
   df_temp<-subset(df_temp,!grepl("[a-z]:",df_temp$`HDI Rank (2018)`))
   df_temp$`HDI Rank (2018)`<-as.numeric(df_temp$`HDI Rank (2018)`)
   df_temp<-df_temp[1:(nrow(df_temp)-17),]
   df_temp$var<-dttyp[i]
   if (i==1) {
     df_ed<-df_temp[!sapply(df_temp, function(x) all(is.na(x)))]}
   else {df_ed<-bind_rows(df_ed,df_temp[!sapply(
     df_temp, function(x) all(is.na(x)))])}
   
}

df_ed<-df_ed[,!grepl("\\.[0-9]",colnames(df_ed))]

```

Once the files are read in correctly we create two separate forms of the education data. One of which is with each variable as a column, and year as its own singular column, utilizing the pivot longer and wider functions. The other method leaves relevant education variables in a singular column, and another column corresponding to their values.  While the second format is not necessarily tidy, this format is kept for future use in analysis.

```{r}
df_ed_lg<-df_ed %>%
            pivot_longer(cols=!c("HDI Rank (2018)","Country","var"),names_to="Year",values_to="HDI") 
df_ed_lg$Year<-as.numeric(df_ed_lg$Year)
df_ed_td<-df_ed_lg %>%
            pivot_wider(names_from="var",values_from="HDI")

df_ed_td %>% write_csv('./Education/Education_Tidy.csv')

```



```{r}

HS <-read.csv("./HumanSecurity/HumanSecurity.Final.Tidy.csv")

full_join(inequality_df,
          full_join(df_ed_td,HS,by=c("Country"="Country","Year"="Year")),
          by=c("Country"="Country","Year"="Year")) %>%
  write_csv('./FinalData/All.csv')


```
