==> 04-missing.Rmd <==
# Missing values

```{r, warning=FALSE,message=FALSE}
library(tidyverse)
library(readxl)
library(readr)
library(stringr)
library(ggplot2)
library(GGally)
library(mi)
```

## Inequality Missing Data


```{r  warning = FALSE,  message = FALSE} 
inequality_df <- read_csv("./Inequality/inequality_tidy.csv")
```

Below is a bar chart showing the proportion of missing data for every Inequality feature.

```{r fig.width = 8}
prop_missing_features <- data.frame(var = names(colMeans(is.na(inequality_df))), 
            prop_missing = unname(colMeans(is.na(inequality_df)))) %>%
          filter(prop_missing > 0) %>%
          filter(!var %in% c('Country', 'HDI Rank (2018)'))

prop_missing_features %>%
    ggplot(aes(x = fct_reorder(var, prop_missing, .desc =FALSE), y = prop_missing)) +
    geom_col() +
    coord_flip() +
    xlab("") +
    ylab("Proportion of Data Missing") +
    ggtitle("Proportion of Data Missing for Each Inequality Feature")
```


The above shows that for 7 of the 9 Inequality features, 20-25% of entries are missing.

As Inequality data includes only 9 years and 9 features, a heatmap is an appropriate tool to visualize combinations of missing data between `Year` and the different Inequality features, with the goal of identifying trends in proportion of missing data for specific features across years (and vice-versa). 

```{r fig.height = 6, fig.width = 8}

prop_missing_year<- inequality_df %>% 
  group_by(Year) %>% 
  summarise_all(~ sum(is.na(.))/n()) %>%
  ungroup() %>% 
  select(-Country, -'HDI Rank (2018)') %>% 
  pivot_longer(-Year, names_to = 'features' , values_to = 'year_prop_missing') 

prop_missing_year%>%
  ggplot(aes(x = factor(Year), y = fct_reorder(features, -year_prop_missing, mean))) +
  geom_tile(aes(fill = year_prop_missing)) +
  scale_fill_continuous(high = "#132B43", low = "#56B1F7") +
  xlab("Year") +
  ylab('') +
  ggtitle('Heat Map of Missing Data for Inequality \nFeature - Year Combination') +
  labs(fill='Proportion Missing') 
  
```


The above heatmap is drawn such that the feature with the largest proportion of missing data is at the bottom. From the heatmap, it seems 2011 and 2012 show the highest rates of missing data for every feature except "Inequality in life expectancy (%)" and "Inequality-adjusted life expectancy index". However, note that the proportion of missing data peaks at around 0.3 and most feature-year combinations show a proportion of missing data below 0.2, which is manageable.  

Below is the aggregate sum of number of missing entries over all features for each country, filtering only the countries with at least one missing entry.  

```{r fig.height = 12, fig.width = 8}

prop_missing_countries<- inequality_df %>% 
  group_by(Country) %>% 
  summarise_all(~ sum(is.na(.))/n()) %>%
  ungroup()

prop_missing_countries <- data.frame(Country = prop_missing_countries$Country, 
                                     country_prop_missing =  rowMeans(prop_missing_countries[,4:12])) %>% 
                          filter(country_prop_missing > 0) 

prop_missing_countries %>% 
    ggplot(aes(x = fct_reorder(Country, country_prop_missing, .desc =FALSE), y = country_prop_missing)) +
    geom_col() +
    coord_flip() + 
    xlab('') +
    ylab('Proportion of Data Missing') +
    ggtitle('Proportion of Missing Inequality Data for Each Country')
    
```


100% of the data is missing for Saint Kitts and Nevis, Liechtenstein, and Dominica. In addition, more than 87.5% of the data is missing for Nauru, Marshall Islands, Andorra, and Tuvalu. All of these countries are extremely small (most are islands); it is possible that Inequality data is not collected nationally and/or data collection is not prioritized by the relevant international organizations.  

A parallel coordinate plot was drawn to assess the presence of a relationship between the proportion of missing data and each Inequality feature, with each country represented by one line.

```{r, fig.height = 8, fig.width=10}

inequality_df[,c(1,2,3,4,12,6, 7,10, 5, 9, 11, 8)] %>% 
  group_by(Country) %>% 
  summarise_all(~ sum(is.na(.))/n()) %>%
  ungroup() %>% 
  select(-Year, -'HDI Rank (2018)', -Country) %>%
  rename("Coefficient\nof Human\nInequality" = "Coefficient of Human Inequality",
        "Overall loss\n in HDI\n due to \ninequality\n(%)" = "Overall loss in HDI due to inequality (%)",
        "Inequality\nadjusted \nHDI\n(IHDI)" = "Inequality-adjusted HDI (IHDI)",
        "Inequality\nadjusted\nincome\nindex" = "Inequality-adjusted income index",
        "Inequality\nin\nincome\n(%)" = "Inequality in income (%)",
        "Inequality\nadjusted\neducation\nindex" = "Inequality-adjusted education index",
        "Inequality in \neducation\n(%)" = "Inequality in education (%)",
        "Inequality\nin life\nexpectancy\n(%)" = "Inequality in life expectancy (%)",
        "Inequality\nadjusted\nlife\nexpectancy\nindex" = "Inequality-adjusted life expectancy index"
         ) %>%
  ggparcoord(alpha = 0.2,
             scale = 'globalminmax',
             splineFactor = 0)  +
  theme(axis.text.x = element_text(angle = 0)) +
  xlab("") +
  ylab("") + 
  ggtitle("Inequality Proportion of Missing Data Parallel Coordinate Plot") +
        theme(plot.title = element_text(size = 18),
              axis.text.x = element_text(size = 12)) 

```


* Most countries appear to not be missing data for Inequality-adjusted life expectancy index and Inequality in life expectancy (%), and when they do, both features are missing in similar proportions. 

* For remaining features, at intervals of around 0.1 from 0 to 1, there are clusters of countries missing that proportion of data for each feature. 

* There are a small number of countries missing data for most to all features (seen as a flat line at the top of the plot). These may be the smaller countries reflected in the bar chart above.

* For Coefficient of Human Inequality, Overall Loss in HDI due to Inequality (%), and Inequality-Adjusted HDI (IHDI), countries generally miss the same proportion of data for all 3 features, suggesting they may be related. 

* In contrast, countries generally miss the same proportion of data in Inequality-Adjusted Income Index as in Inequality in Income (%), suggesting they may be related. In addition, countries are generally missing as much or less data for these features as they are for Coefficient of Human Inequality, Overall Loss in HDI due to Inequality (%), and Inequality-Adjusted HDI (IHDI).

* For Inequality-Adjusted Education Index and Inequality in Education (%), countries generally miss the same proportion of data; again, they may be related.



## Human Security Missing Data

The following abbreviations will be used for variables examined in this section: 

* Homicide Rate (per 100,000 people) - Homicide.Rate

* Prison Population (per 100,000 people) - Prison.popn

* Suicide Rate (Male) (per 100,000 people) - Suicide.rate.M

* Suicide Rate (Female) (per 100,000 people) - Suicide.rate.F

A significant proportion of data in the Human Security dataset is missing.  Largely, this seems to be a product of uneven tracking between variables and geographic complications.

```{r, echo=F, warning=FALSE,message=FALSE}
Human.Security <- read_csv("./HumanSecurity/HumanSecurity.Final.Tidy.csv")
Human.Security <- subset(Human.Security, select = -c(3))
Human.Security <- Human.Security[rowSums(is.na(Human.Security[,3:6])) != ncol(Human.Security[,3:6]),]
Homicide.rate <- read_csv("./HumanSecurity/Homicide.tidy.csv")
Homicide.rate <- Homicide.rate[,2:14]
Homicide.rate.1 <- read.csv("./HumanSecurity/Homicide.tidy.csv")
Homicide.rate.1 <- Homicide.rate.1[,2:14]
names(Homicide.rate.1)[2:13] <- c("1990", "1995", "2000", "2005", "2010", "2011", "2012", "2013", "2014", "2015", "2016", "2017")
Suicide.rate.F <- read_csv("./HumanSecurity/SR.F.tidy.csv")
Suicide.rate.F <- Suicide.rate.F[,2:6]
Suicide.rate.M <- read_csv("./HumanSecurity/SR.M.tidy.csv")
Suicide.rate.M <- Suicide.rate.M[,2:6]
Prison.popn <- read.csv("./HumanSecurity/Prison.population.csv")
Prison.popn <- subset(Prison.popn, select = -c(X, X.1, X.2, X.3, X.4, X.5, X.6, X.7, X.8, X.9))
names(Prison.popn)[2:11] <- c("2003", "2005", "2010", "2011", "2012", "2013", "2014", "2015", "2016", "2017")
dcol <- 2
for (i in 2:ncol(Prison.popn)) {
  Prison.popn[,dcol] <- as.numeric(Prison.popn[,dcol])
  dcol <- dcol +1
}
```

The dataset shows uneven tracking between variables. For example, Homicide Rate is recorded in 1990, 1995, 2000, 2005, and yearly from 2010 to 2017. Suicide Rate for both males and females, by contrast, was recorded only in the years 2000, 2010, 2015, and 2016.

Below are the years in the Homicide Rate dataset for which data were collected:

```{r}
colnames(Homicide.rate)[2:13]
```

By contrast, Suicide Rate data for both males and females were only collected for the following years:

```{r}
colnames(Suicide.rate.F)[2:5]
```

The difference in frequency of data collection between these datasets results in a merged dataframe with a high proportion of missing values. 

Specifically, the consistency of the Prison Population and Homicide Rate data introduces years into the dataframe for which Suicide Rate shows no recorded data. As a result, Suicide Rate (M) and Suicide Rate (F) are both missing values over more than 60% of the dataset, as shown in the graph below.

```{r}
prop_missing_features <- data.frame(var = names(colMeans(is.na(Human.Security))), 
            prop_missing = unname(colMeans(is.na(Human.Security)))) %>%
          filter(prop_missing > 0)  

prop_missing_features %>%
    ggplot(aes(x = fct_reorder(var, prop_missing, .desc =FALSE), y = prop_missing)) +
    geom_col() +
    coord_flip() +
    xlab(NULL) + 
    ylab("Proportion of Data Missing") +
    ggtitle("Missing Data", subtitle = "Including all years") + 
    ylim(c(0,.9))
```

Alternately, the graph below gives a more accurate picture of the proportion of missing data for each variable.


```{r, echo=F}
hs_missing <- as.data.frame(matrix(nrow = 4, ncol = 2))
colnames(hs_missing) <- c("factor", "prop.missing")
hs_missing[,1] <- c("Homicide.rate", "Prison.popn", "Suicide.F", "Suicide.M")
hs_missing[1,2] <- mean(as.matrix(is.na(Homicide.rate)))
hs_missing[2,2] <- mean(as.matrix(is.na(Prison.popn)))
hs_missing[3,2] <- mean(as.matrix(is.na(Suicide.rate.F)))
hs_missing[4,2] <- mean(as.matrix(is.na(Suicide.rate.M)))

hs_missing %>%
    ggplot(aes(x = fct_reorder(factor, prop.missing, .desc =FALSE), y = prop.missing)) +
    geom_col() +
    coord_flip() +
    xlab(NULL) + 
    ylab("Proportion of Data Missing") +
    ggtitle("Missing Data", subtitle = "Including only collected years for each variable") +
    ylim(c(0,.9))
```

This chart reflects the average proportion of missing data for each variable (across all recorded years and countries). Calculating these proportions separately excludes the empty values introduced by the unequal frequency of data collection between variables. The variables with the greatest proportion of missing data (Prison Population and Homicide Rate) show between 25% and 35% missing data, respectively.

The following heatmap shows missing data split by variable.  Importantly, the map reflects the entire Human Security dataset, again highlighting trends in missing data due to uneven frequency of data collection.

```{r}
prop_missing_year<- Human.Security %>% 
  group_by(Year) %>% 
  summarise_all(~ sum(is.na(.))/n()) %>%
  ungroup() %>% 
  select(-Country) %>% 
  pivot_longer(-Year, names_to = 'features' , values_to = 'year_prop_missing') 

prop_missing_year%>%
  ggplot(aes(x = factor(Year), y = fct_reorder(features, -year_prop_missing, mean))) +
  geom_tile(aes(fill = year_prop_missing)) +
  scale_x_discrete(expand=c(0,0)) +
  scale_fill_continuous(high = "#132B43", low = "#56B1F7") +
  xlab("Year") + 
  ylab(NULL) +
  ggtitle("Missing Human Security Data by Variable") +
  labs(fill = 'Proportion Missing')
```

Clearly, Suicide Rate (M) and Suicide Rate (F) are missing the greatest amount of data, though there seems to be relatively little missing data in the reported years.  Both Homicide Rate and Prison Population are consistently reported between 2005 and 2017, with slight variations in missing data. Homicide Rate shows no data reported for the year 2003, while Prison Population shows no data before 2003.

A number of unique problems may be incurred in the collection of international data, as discussed briefly in the Data chapter. Certain countries or regions may have presented challenges for data collection, given that collection processes relied on state institutions and infrastructure.

The Homicide Rate dataset is a particularly demonstrative example.  The final entries in the dataset show countries for which there is no recorded homicide data.

These countries are:

```{r}
print(Homicide.rate$Country[168:192])
```

Of these 24 countries, 20 (Cóte d'Ivoire, Congo, Djibouti, Ethiopia, Gabon, Gambia, Madagascar, Mali, Nigeria, Sudan, Chad, Congo [Dem. Republic], Comoros, Eritrea, Guinea, Equatorial Guinea, Libya, Mauritania, Senegal, and Togo) are in Africa. The remaining four (Lao [People's Dem. Republic], Vanuatu, Micronesia [Fed. States], the Marshall Islands, and Palau) are small island countries in Oceania.  Potential association between geographical region and missing data represents an important direction for further research, and provides an example of possible trends in missing data.


```{r fig.height = 30, fig.width = 8}

prop_missing_countries <- Human.Security %>% 
  group_by(Country) %>% 
  summarise_all(~ sum(is.na(.))/n()) %>%
  ungroup()

prop_missing_countries <- data.frame(Country = prop_missing_countries$Country, 
                                     country_prop_missing =  rowMeans(prop_missing_countries[,3:6])) %>% 
                          filter(country_prop_missing > 0)

prop_missing_countries %>% 
    ggplot(aes(x = fct_reorder(Country, country_prop_missing, .desc =FALSE), y = country_prop_missing)) +
    geom_col() +
    coord_flip() + 
    xlab(NULL) + 
    ylim(c(0,0.9)) +
    ylab("Proportion of Data Missing") +
    ggtitle("Missing Human Security Data by Country")
```

The above graph shows the proportion of missing data among the Human Security variables by country. 

Note that San Marino, Palau, Nauru, and the Marshall Islands are each missing 75% of data.

Due to the large volume of countries included in this data, and their alphabetical organizational scheme, a heatmap is not an appropriate visualization tool to examine patterns of missing data by country.  


## Education Missing Data


```{r warning = FALSE,  message = FALSE}

df_ed_td<-read_csv('./Education/Education_Tidy.csv')
df_ed_lg<-read_csv('./Education/Education_LG.csv')
df<-df_ed_td[c(2,4:25)] %>% group_by(Country) %>% select(everything()) %>%
     summarise_all(funs(sum(is.na(.)))) 
df$TotalMiss<-rowSums(df[,2:23])
df$PerTot<-df$TotalMiss*100/sum(df$TotalMiss)

```


There are `r format(sum(df$TotalMiss),scientific=F)` missing observations across all Education variables, years, and countries. The following bar chart displays an interesting yet confusing pattern. All countries appear to have between `r round(min(df$TotalMiss),0)` and  `r round(max(df$TotalMiss),0)`, corresponding to `r round(min(df$PerTot),2)` and `r round(max(df$PerTot),2)` percent  total missing entries. Keeping in mind the data spans 29 years and 22 variables, these measurements seem reasonable, though interestingly there seems to be no apparent overarching pattern between missing observations. Some smaller countries, however, do seem more likely to have missing data.

```{r,fig.height=25,fig.width=10}
ggplot(df,aes(y=fct_reorder(Country,PerTot),x=PerTot)) +
geom_col()+ylab("")+xlab("Proportion")+ggtitle("Education: Proportion Missing by Country")+
  theme(plot.title = element_text(hjust = 0.5),text = element_text(size=13))
```


For this analysis, we want the variables with the least missing values across Countries. While the full chart is too large and messy to display, plotting Country versus Variable displays some interesting insights. Overall, some variables can already be narrowed down. For example, variables that show little to no missing data across all countries (such as Education Index, or Mean Years of Schooling) are noteworthy. Among other interesting notes, the former Yugoslav Republic of Macedonia seems to be missing all observations--which is consistent with the original source data. After a little digging, it seems the country's inclusion is an oversight; "Yugoslav Republic of Macedonia" was replaced by "North Macedonia," and was likely a left over data point. Very few countries reported the percent of students without access to internet, and of the countries that reported, most seem to be highly developed. The percent of the population with at least some secondary education, by contrast, seems to be consistently reported.

```{r,fig.height=35,fig.width=20,eval=FALSE}

ggplot(df[,1:23] %>%
     pivot_longer(cols=!c("Country"),names_to="Vari",values_to="Val") ,aes(y=Country,x=Vari,fill=Val))+
  geom_tile()+ylab("")+xlab("Variable")+ggtitle("Education: Missing by variable and Country")+
  theme(plot.title = element_text(hjust = 0.5),text = element_text(size=17),axis.text.x=element_text(angle=90)) +
  scale_fill_continuous(high = "#132B43", low = "#56B1F7")

```


Breaking down this analysis a little further, the lack of data across countries on a consistent basis before 2010 suggests that analysis should focus on 2010 and later. The data appears to have been released in 5-year increments. 

The following heatmap was included to display yearly patterns across countries, though its messiness may hinder its readability to a broader audience. We debated whether to keep this chart along with the previous hidden one. But it was argued that the maps allows one to confirm whether there are patterns across countries by year (and variable), and there quite clearly is. Hence, while messy it allows one to observe this important piece of information. It is also quite important to determine missingness across country, the lack of which would hinder one's ability to develop a complete analysis of missings. For example--what if there were several countries with missings across year or variable? One would not be able to observe this clearly like was done for The former Yugoslav Republic of Macedonia. Hence it can be argued that for EDA the charts present important information.

```{r,fig.height=40,fig.width=25}

df<-df_ed_lg %>% pivot_wider(names_from="Year",values_from="HDI")
df_n<-df[c(2,4:32)] %>% group_by(Country) %>% select(everything()) %>%
     summarise_all(funs(sum(is.na(.)))) 
ggplot(df_n %>%
     pivot_longer(cols=!c("Country"),names_to="Year",values_to="Val") ,aes(y=Country,x=Year,fill=Val))+
  geom_tile()+ylab("")+xlab("Year")+ggtitle("Education: Missing by Year and Country")+
  theme(plot.title = element_text(hjust = 0.5),text = element_text(size=18,face="bold"),axis.text.x=element_text(angle=90)) +
  scale_fill_continuous(high = "#132B43", low = "#56B1F7")


```

Interestingly, breaking down the missing variables by year (below) confirms what was seen previously: the most reliable variables seem to be Education Index (EdInd), Mean Years of Schooling (MYSchl), and Expected Years of Schooling (ExpSch; according to the amount of missing data across years), as well as the variable for "population with some secondary education" (PopSS25U). The following heatmap also suggests data reporting may have been lagged, as 2018 data seems to be missing for variables which have otherwise reported more consistent data.

```{r,fig.height=25,fig.width=20}
df<-df_ed_td[c(3:25)] %>% group_by(Year) %>% select(everything()) %>%
     summarise_all(funs(sum(is.na(.)))) 
df$Year<-factor(df$Year)
ggplot(df %>%
     pivot_longer(cols=!c("Year"),names_to="Variable",values_to="Val") ,aes(y=Year,x=Variable,fill=Val))+
  geom_tile()+xlab("Variable")+ylab("Year")+ggtitle("Education: Missing by Year and Variable")+
  theme(plot.title = element_text(hjust = 0.5),text = element_text(size=17, face="bold"),axis.text.x=element_text(angle=90))+
  scale_fill_continuous(high = "#132B43", low = "#56B1F7")

```
