==> 05-results.Rmd <==
# Results

## Variable Selection 


```{r,warning=FALSE,message=FALSE}

library(tidyverse)
library(readxl)
library(readr)
library(stringr)
library(ggplot2)
library(GGally)

```

### Inequality Variable Selection

From Section 4.0 Missing Data, it was noted (from the Inequality parallel coordinate plot) that the following sets of features have the same proportion of missing data for nearly every country, indicating they may be related: 

* Coefficient of Human Inequality, Overall Loss in HDI due to Inequality (%), and Inequality-adjusted HDI (IHDI)

* Inequality-Adjusted Income Index and Inequality in Income (%)

* Inequality-Adjusted Education Index and Inequality in Education (%)

* Inequality-Adjusted Life Expectancy Index and Inequality in Life Expectancy (%)

A scatterplot matrix of all 9 features without the missing data is shown below, so that pairwise correlations between all features can be examined. This analysis begins with the feature groups listed above, with strong relationships in their missing data, then extends to the remaining feature variable pairs. 


```{r,fig.height=12, fig.width = 12, warning = FALSE ,  message = FALSE}
inequality_df <- read_csv("./Inequality/inequality_tidy.csv")
inequality_df_copy <- inequality_df
names(inequality_df_copy) <- gsub("-", " ", names(inequality_df_copy))
ggpairs(inequality_df_copy[,c(4,12,6, 7,10, 5, 9, 11, 8)], progress = FALSE,
        upper = list(continuous = wrap("cor", size = 4)),
        lower = list(continuous = wrap('points', alpha = 0.1)),
        labeller = label_wrap_gen(10))+
      ggtitle("Pairwise Correlation Scatterplot Matrix", subtitle = 'Inequality') + 
      theme(plot.title = element_text(size = 24),
            plot.subtitle = element_text(size = 20),
            strip.text = element_text(size = 12)) 

```



First, note the feature groups where missing data appear to be in equal proportions for most countries:

* Coefficient of Human Inequality, Overall Loss in HDI due to Inequality (%), and Inequality-Adjusted HDI (IHDI): The three pairwise correlation coefficients between these three features are 0.999, -0.954, and -0.950, showing  these three feature are highly correlated when the data is not missing.

* Inequality-Adjusted Income Index and Inequality in Income (%): Correlation coefficient of -0.644. Despite countries missing data for both feature consistently, the two features do not appear to have as strong of a correlation as other cases discussed. 

* Inequality-Adjusted Education Index and Inequality in Education (%): The correlation coefficient is -0.942, showing that these two features are highly correlated when the data is not missing.

* Inequality-Adjusted Life Expectancy Index and Inequality in Life Expectancy (%): The correlation coefficient is -0.976, showing that these two features are highly correlated when the data is not missing.

On the basis of the results above, the following features can be dropped without significant loss of information:

* Overall Loss in HDI due to Inequality (%) - due to strong correlation with Coefficient of Human Inequality and the fact both share similar missing patterns. 

* Inequality-Adjusted HDI (IHDI) - highly correlated with multiple features, especially Coefficient of Human Inequality and Overall Loss in HDI due to Inequality (%) (while also sharing the same missing pattern as the latter two).

* Inequality-Adjusted Education Index - highly correlated with Inequality in Education (%) when data is present and in terms of missing pattern.  

* Inequality-Adjusted Life Expectancy Index - highly correlated with Inequality in Life Expectancy (%) when data is present and in terms of missing pattern.  

Looking at the pairplot of the 5 remaining variables, note that even though the variables do not share the same missing patterns, Coefficient of Human Inequality is strongly correlated with 3 of the remaining features (absolute correlation coefficient of ~0.9). Therefore, this feature can also be dropped. As a result, the remaining 'Inequality' features kept for analysis with other indicators are:

* Inequality-Adjusted Income Index

* Inequality in Income (%)

* Inequality in Education (%)

* Inequality in Life Expectancy (%)



### Education Variable Selection

Initially we pulled only the variables with the least missing data. We then ran ggpairs across these variables to determine if they could be further narrowed in selection. Given that the variables are all relatively highly correlated, it seems Education Index (EdInd) provides the clearest comparison in analysis of correlations across human development topics.

```{r warning = FALSE,  message = FALSE,fig.height=10,fig.width=10}

df_ed<-read_csv('./Education/Education_Tidy.csv')
df_ed<-df_ed[,c("HDI Rank (2018)","Country","Year",
        "EdInd","ExpSch","MYSch","PopSS25U")]
ggpairs(df_ed[,c("EdInd","ExpSch","MYSch","PopSS25U")])+
ggtitle("Pairwise Correlation Scatterplot Matrix", subtitle = "Education")+
theme(text = element_text(size=17))
```

Next, in determining how to handle the 'Year' variable, we examine the yearly breakdown after 2010 (as the Inequality dataset only spans 2010-2018). From here, each year appears highly correlated (.985 to .999), suggesting years do not need to be examined individually, and can be aggregated with a simple average.

```{r warning = FALSE,  message = FALSE,fig.height=15,fig.width=15}
df<-df_ed[df_ed$Year>=2010,c('Country','Year','EdInd')] %>%
  pivot_wider(names_from='Year',values_from="EdInd")
ggpairs(df[,c('2010','2011','2012','2013','2014','2015','2016','2017','2018')])+
theme(text = element_text(size=17))+
ggtitle("Pairwise Correlation Scatterplot Matrix", subtitle = "Education Index by Year")

```


### Human Security Variable Selection

The following abbreviations will be used for variables examined in this section: 

* Homicide Rate (per 100,000 people) - Homicide.Rate

* Prison Population (per 100,000 people) - Prison.popn

* Suicide Rate (Male) (per 100,000 people) - Suicide.rate.M

* Suicide Rate (Female) (per 100,000 people) - Suicide.rate.F

As noted in section 4, the Suicide Rate data are collected sporadically.  While these datasets internally show the lowest proportion of missing data, they only reflect data collected for four years, (2000, 2010, 2015, and 2016) which are not consistently spaced across time.

The other datasets in this section (Homicide rate and Prison population) show a higher proportion of missing data, but reflect a larger number of consistently sampled years.

Finally, as the Refugees by Country of Origin and Homelessness due to Natural Disasters data are unavailable, they cannot be included in this analysis.

The following graph shows the pairwise correlation between all variables in the dataset, with the exception of Refugees by Country of Origin and Homelessness due to Natural Disaster.


```{r,fig.height=8, fig.width = 8, warning = FALSE ,  message = FALSE}
Human.Security <- read_csv("./HumanSecurity/HumanSecurity.Final.Tidy.csv")
Human.Security <- subset(Human.Security, select = -c(3))
Human.Security <- Human.Security[rowSums(is.na(Human.Security[,3:6])) != ncol(Human.Security[,3:6]),]
ggpairs(Human.Security[,c(3:6)], progress = FALSE,
        upper = list(continuous = wrap("cor", size = 4)),
        labeller = label_wrap_gen(10)) + 
      theme(strip.text = element_text(size = 8)) + 
      ggtitle("Pairwise Correlation Scatterplot Matrix", subtitle = "Human Security")

```

With the exception of male and female suicide rate (0.552), none of the data show correlation coefficients above 0.5 or below -0.5, indicating they are not highly correlated. 

As a result, the 'Human Security' variables selected for this study are the following: 

* Homicide Rate (per 100,000 people)

* Prison Population (per 100,000 people)


## Education versus Human Security (2010 - 2018)

For the actual analysis, we compare the average across 'Education Index' (EdInd) against the 'Human Security' variables:

* Homicide Rate (per 100,000 people) (Homicide.rate)

* Prison Population (per 100,000 people) (Prison.popn)

```{r warning=FALSE, message=FALSE}
All<-read_csv('./FinalData/All.csv')
All<-All[All$Year>=2010,c("Country","EdInd",  
                          'Homicide.rate','Prison.popn')]
EdHS<-All %>% group_by(Country) %>% select(everything()) %>%
  summarise_all(~mean(., na.rm = TRUE)) 

```

Here it appears that a high Education Index is associated with low Homicide Rate and all levels of Prison Population. Education and Homicides seem to have a direct relation with higher education implying lower homicide rates-- a somewhat inverse relationship. Education and Prison Population seem to show a similar relation, though a high Education Index seems to be associated more with a median level of Prison Population suggesting a parallel relationship. Notably, several countries with high education levels show higher prison populations.

```{r}
ggparcoord(EdHS,scale = 'uniminmax',splineFactor = 5)  +
    theme(axis.text.x = element_text(angle = 0)) +
    xlab("") +
    ylab("") +
    ggtitle("Human Security versus Education Parallel Coordinates Plot")

```


The following scatterplot matrix seems to show the relationship clearly. While higher levels of education are still associated with low Homicide Rates, restricting the analysis to lower levels of education shows an associated jump in high Homicide outliers. As expected the sign here is negative, confirming the likelihood of a negative relation between the variables.

The relationship between Education and Prison Population, though is somewhat stronger and more direct, as seen in the matrix, with a correlation of 0.3. Thus, higher levels of education seem to imply higher levels of imprisonment.

```{r warning=FALSE, message=FALSE}

EdHS %>%
  ggpairs(columns = c("EdInd",  
                          "Homicide.rate","Prison.popn"))+
  ggtitle("Pairwise Correlation Scatterplot Matrix", subtitle = "Education Index Against Human Security")


```


## Inequality versus Education (2010 - 2018)

This section details the analysis of Inequality versus Education indicators for the selected features:

Inequality

* Inequality-Adjusted Income Index

* Inequality in Income (%)

* Inequality in Education (%)

* Inequality in Life Expectancy (%)

Education

* Education Index

* Mean Years of Schooling (Years)


```{r, warning = FALSE, message = FALSE}
education_df <- read_csv("./Education/Education_Tidy.csv") %>%
                  filter(Year %in% c(2010:2018)) %>%
                  select(Country, Year, EdInd, MYSch)

inequal_ed <- inequality_df %>% 
  select(Country, Year, 
         'Inequality-adjusted income index', 
         'Inequality in income (%)', 
         'Inequality in education (%)',
         'Inequality in life expectancy (%)') %>%
  full_join(education_df, by = c('Country', 'Year'))

temp <- gsub(" ", "_", names(inequal_ed))
temp <- gsub("-", "_", temp)
temp <- gsub("\\(%\\)", "perc", temp)
names(inequal_ed) <-  temp

```


First, all Inequality variables were compared with all Education variables in a scatterplot faceted by year from 2010 to 2018, as those were the only years with available Inequality data. It was noted that "Year" did not separate the data, as illustrated in the a sample facet plot below for Mean Year of Schooling versus Inequality in Life Expectancy (%). This was the case for every combination of variables. 

```{r, warning = FALSE, fig.height= 8, fig.width= 8}
inequal_ed %>%
  ggplot(aes(x = Inequality_in_life_expectancy_perc, y = MYSch)) +
  geom_point() +
  facet_wrap(~Year) +
  xlab('Inequality in Life Expectancy (%)') +
  ylab('Mean Years of Schooling (Years)') +
  ggtitle('Sample Plot of Inequality in Life Expectancy versus Mean Years of \nSchooling Faceted by Years') +
        theme(plot.title = element_text(size = 16))
```


As there are 6 features in total, all of which are continuous, a parallel coordinate plot is the most appropriate visualization tool for this type of data. Each of the six features was grouped by country and averaged over all years from 2010 to 2018, ignoring NA's. The rationale is as follows:

* It was shown from the facet plots previously that "Year" did not distinguish any trends, consistent with the trend discovered in "Education Variable Selection" and extends to all other feature pairs between indicators in subsequent analysis.   

* If every Country-Year combination is a line on the parallel coordinate plot, then it would require that none of the six features have NA's. This is too strong and may lead to underplotting given the amount of missing data in this dataset. This also ensures that every country also shows up only once or not at all. 


```{r, fig.height = 7, fig.width=10}
inequal_ed[, c(1,2,6,4,3,5,7,8)] %>%
  group_by(Country) %>%
  summarise_all(~mean(., na.rm = TRUE)) %>%
  ungroup() %>%
  rename('Inequality in\nIncome (%)' = Inequality_in_income_perc, 
         "Inequality in \nEducation (%)" = Inequality_in_education_perc,
         "Inequality\nAdjusted \nIncome Index" = Inequality_adjusted_income_index,
         "Inequality in \nLife Expectancy\n(%)" = Inequality_in_life_expectancy_perc,
         'Education Index ' = EdInd,
         'Mean Years\nof Schooling\n(Years)' = MYSch) %>% 
  select(-Country, -Year) %>% 
  ggparcoord(alpha = 0.25,
               scale = 'uniminmax',
               splineFactor = 0)  +
    theme(axis.text.x = element_text(angle = 0)) +
    xlab("") +
    ylab("") +
    ggtitle("Inequality versus Education Parallel Coordinate Plot") +
        theme(plot.title = element_text(size = 20),
              axis.text.x = element_text(size = 14)) 
```

Observations:

* The two Education variables, Education Index and Mean Years of Schooling, are positively correlated. Therefore, analysis for this section will be made with Education Index only because the same conclusions will apply to Mean Years of Schooling.

* There is a large cluster of countries (Cluster 1) with low Inequality in Income (%), mid to high Inequality Adjusted Income Index, low Inequality in Education (%), and mid to high Education Index. 

* For Cluster 1, there's a visible negative correlation between Inequality in Education (%) and Education Index. 

* As Inequality-Adjusted Income Index is visibly negatively correlated with Inequality in Education (%) in Cluster 1, it can be inferred that the former is positively correlated with Education Index. 

* There is another cluster of countries (Cluster 2) exhibiting the opposite trend. This cluster of countries has low to mid Inequality Adjusted Income Index, high Inequality in Education (%), and low-mid Education Index.

* For countries in Cluster 2, there is no obvious negative correlation between Inequality-Adjusted Income Index - Education Index and Inequality in Education (%) - Education Index like in Cluster 1.

* Cluster 1 has a visible right skew in Inequality in Education (%) and a slight left skew in Education Index, while countries in Cluster 2 exhibit no significant skew for these two features compared to Cluster 1. 



## Inequality versus Human Security (2010 - 2018)

This section details the analysis of Inequality versus Human Security given the following variables:

Inequality

* Inequality-Adjusted Income Index

* Inequality in Income (%)

* Inequality in Education (%)

* Inequality in Life Expectancy (%)

Human Security

* Homicide Rate (per 100,000 people)

* Prison Population (per 100,000 people)


```{r, warning = FALSE, message = FALSE}
HSINEQ_df <- read_csv("./FinalData/All.csv") %>%
                  filter(Year %in% c(2010:2018)) %>%
                  select(Country, Year, 'Inequality-adjusted income index', 'Inequality in income (%)', 'Inequality in education (%)', 'Inequality in life expectancy (%)', 'Homicide.rate', 'Prison.popn')

```


First, Inequality variables were contrasted against Human Security variables in a scatterplot, faceted by year from 2010 to 2018 (again, the only years Inequality data were collected).

```{r, fig.height=8, fig.width=8, warning = FALSE}

hs.pairs <- HSINEQ_df %>%
  group_by(Country) %>%
  summarise_all(~mean(., na.rm = TRUE)) %>%
  ungroup() %>%
  rename("Inequality in \nEducation (%)" = `Inequality in education (%)`,
         "Inequality Adjusted \nIncome Index" = `Inequality-adjusted income index`,
         "Inequality in \nLife Expectancy (%)" = `Inequality in life expectancy (%)`,
         "Prison Population" = Prison.popn,
         "Homicide Rate" = Homicide.rate) %>% 
  select(-Country, -Year)
  
ggpairs(hs.pairs, progress = FALSE,
        upper = list(continuous = wrap("cor", size = 4)),
        labeller = label_wrap_gen(10)) + 
      theme(strip.text = element_text(size = 8)) + 
      ggtitle("Pairwise Correlation Scatterplot Matrix", subtitle = "Inequality versus Human Security")
```

* The association between the Inequality variables (Inequality-Adjusted Income Index, Inequality in Income, Inequality in Education, and Inequality in Life Expectancy) is discussed in the Inequality variable selection, above.

* Similarly, the association between the Human Security variables (Homicide Rate and Prison Population) are discussed in the Human Security variable selection, above.

* Prison Population seems to show low positive correlations with two Inequality variables: Inequality-Adjusted Income Index and Inequality in Income (with correlation coefficients of .156 and .171, respectively).  Interestingly, Prison Population shows negative correlations with both Inequality in Education and Inequality in Life Expectancy (-0.349 and -0.334, respectively).

* Homicide Rate shows low positive correlations with Inequality in Education and Inequality in Life Expectancy (with correlation coefficients of 0.127 and 0.142).

* Homicide Rate additionally shows correlations with Inequality-Adjusted Income Index, with a negative correlation of -0.324, and Inequality in Income, with a positive correlation of 0.488.

The parallel coordinate plot, below, maps some of these associations.

```{r, fig.height = 7, fig.width=10}
HSINEQ_df[, c(1,2,5,8,3,6,7)] %>%
  group_by(Country) %>%
  summarise_all(~mean(., na.rm = TRUE)) %>%
  ungroup() %>%
  rename("Inequality in \nEducation (%)" = `Inequality in education (%)`,
         "Homicide Rate" = Homicide.rate,
         "Inequality Adjusted \nIncome Index" = `Inequality-adjusted income index`,
         "Prison Population" = Prison.popn,
         "Inequality in \nLife Expectancy (%)" = `Inequality in life expectancy (%)`) %>% 
  select(-Country, -Year) %>% 
  ggparcoord(alpha = 0.25,
               scale = 'uniminmax',
               splineFactor = 0)  +
    xlab("") +
    ylab("") +
    ggtitle("Inequality versus Human Security Parallel Coordinate Plot") +
    theme(plot.title = element_text(size = 20),
          axis.text.x = element_text(size = 14)) 
```
The above graph reorders variables to visualize new trends.

There is a visual cluster of countries marked by a negative relationship between Inequality in Education and Prison Population, followed by a positive relationship between Prison Population and Inequality-Adjusted Income Index, and a negative relationship between Inequality-Adjusted Income Index and Inequality in Life Expectancy.  Some of these countries continue to have what appears to be a direct relationship between Inequality in Life Expectancy and Homicide Rate.

Within this cluster, countries demonstrate low Prison Population, high Inequality-Adjusted Income, and low Inequality in Life Expectancy. Some additionally demonstrate low Homicide Rate.


For our interactive component, we created a .csv file containing the relevant variables from the above analysis, which was uploaded into Github gist and read into our interactive plot in the following section. For countries were data was available, we had one observation per country, averaged over years 2010 - 2018 (ignoring NAs), for each relevant feature.

```{r, warning=FALSE, message=FALSE}

All<-read_csv('./FinalData/All.csv')
All<-All[All$Year>=2010,c("Country","EdInd",  
                          'Homicide.rate','Prison.popn',
                          'Inequality-adjusted income index', 'Inequality in income (%)', 
                          'Inequality in education (%)', 'Inequality in life expectancy (%)')]
AvgAll<-All %>% group_by(Country) %>% select(everything()) %>%
  summarise_all(~mean(., na.rm = TRUE))
AvgAll %>% write_csv("./Interactive/HumDevAll.csv")


```
