==> 06-interactive.Rmd <==
# Interactive component


```{r, eval = FALSE, warning=FALSE, message = FALSE}
library(tidyverse)
library(readr)
read_csv('./Interactive/HumDevAll.csv') %>%
  na.omit() %>%
  write_csv('./Interactive/HumDevAll_narm.csv')
```


<script src="https://d3js.org/d3.v6.js"></script>

<style type="text/css">
  .xlabel{
    font-size: 16px
  }
  .ylabel{
    font-size: 16px
  }
  
  
  #Buttons{
    float:right;

 //   width:10%;
 //   margin-top:150px;
  }

  svg{
  //  float:right;
  //  width: 50%;

  }

  h3{
  //  margin-left:300px;
  }

</style>

<h3>Education - Inequality - Human Security Feature Relationships</h3>
<div id="human_dev">
<div id="Buttons">
  <select id="selectButtonX"></select> 
  <p>X</p>
  <select id="selectButtonY"></select>
  <p>Y</p>
</div>
</div>


<script src="Interactive/human_development.js"></script>

<p></p>
Our interactive component was built upon [a basic plot from the D3.js Graph Gallery](https://www.d3-graph-gallery.com/graph/connectedscatter_select.html). To err on the side of simplicity, "Inequality in Income (%)" was not included in the interactive component since from our analysis, it had the least importance and also had moderate correlation with "Inequality Adjusted Income Index" which is included.
