==> 07-conclusion.Rmd <==
# Conclusion

This study found a number of notable associations between variables, which motivate further research. 

Between Inequality and Human Security, both Homicide Rate and Prison Population showed significant associations.  Homicide Rate demonstrated mid-strength correlations with Inequality-Adjusted Income Index (negative), and Inequality in Income (%) (positive). Prison Population showed mid-strength negative correlations with both Inequality in Education (%) and Inequality in Life Expectancy. Further, there appears to be a cluster of countries that demonstrate low Prison Population, high Inequality-Adjusted Income Index, low Inequality in Life Expectancy, and low Homicide Rate. While outside the scope of this study, identifying the HDI rankings of the countries within this cluster may prove useful in future analysis.  Additionally, future research into potentially confounding variables would assist in identifying policy implications; for example, research investigating possible association between racial or cultural homogeneity and Inequality in Income, Inequality in Life Expectancy, or Prison Population, associations between government infrastructure and Inequality in Income or Prison Population, etc.

In regard to Human Security indices, the effect of Higher Education seems mixed. High Education appears to be associated with higher Prison Population; further research might investigate prison programs and the differences between incarceration and rehabilitation policies.  Additionally, Homicide Rate trends low among highly educated countries, and demonstrates larger gaps between countries with education indices that fall below the international average. Many of the high outliers within the Homicide Rate data show relatively low education rates, though overall this pattern does not imply a strong relationship. Future research  might involve case study analysis of outliers, or a more in-depth analysis of the trends among countries that fall below international averages for education.

Finally, in regard to the Inequality and Education variables, this study found two distinct clusters with relationships across several variables.  The first cluster (Cluster 1) demonstrates low Inequality in Income (%), mid to high Inequality Adjusted Income Index, low Inequality in Education (%), and mid to high Education Index. Within this cluster, Inequality in Education (%) is negatively correlated with Education Index, but Inequality Adjusted Income Index is positively correlated with Education Index.  The second cluster (Cluster 2) shows low to mid Inequality Adjusted Income Index, high Inequality in Education (%), and low-mid Education Index. Within Cluster 2, there is no obvious negative correlation between Inequality Adjusted Income Index and Education Index or between Inequality in Education (%) and Education Index, as seen in Cluster 1. Future research might helpfully explore correlations (i.e. regional) between countries that lie in clusters 1 and 2. Such research might seek to address why countries in clusters 1 and 2, between the same variables, exhibit almost opposite correlation behavior.
